fix(integrity): eliminate hardcoded values, fake passes, and benchmark skew
- Replace hardcoded DRAM and swizzle numbers in src/verify_*.py with authentic analytical and simulation models
- Symmetrize rollback benchmark in benchmarks/benchmark_m1_kv_cache.py without baseline crop pollution
- Fix un-crippled FP16 Tensor Core execution for baseline matmul in tests/test_h6_fused_backward_adamw.py
- Normalize compare_grpo_curves.py by step count and unify throughput calculation across GRPO benchmarks
- Fix TruthfulQA evaluation loophole in benchmarks/benchmark_indepth_honesty.py
- Tighten error gates in tests/test_fused_gemm_correctness.py, test_ellie_custom_t4_kernel.py, and test_h17_fused_int3_gemv.py
- Enforce non-zero exit codes on test failures and explicit [SKIP] logs when CUDA is missing
- Fix acceptance gate condition in scripts/benchmark_t4_speculative.py

+++ tests/test_h6_fused_backward_adamw.py
@@ -175,7 +175,7 @@ def test_h6_correctness_and_benchmark():
 
     start_event.record()
     for _ in range(iters):
-        dW = torch.matmul(dY.t().float(), X.float())
+        dW = torch.matmul(dY.t(), X).float()
         W_master -= lr * weight_decay * W_master
         exp_avg.mul_(beta1).add_(dW, alpha=1.0 - beta1)
         exp_avg_sq.mul_(beta2).addcmul_(dW, dW, value=1.0 - beta2)
